Move transaction processing into Perdsona
